src/src/requirements.txt
# --- requirements.txt ---
# LƯU Ý: PyTorch và Torchvision KHÔNG cài qua file này trên Jetson Nano.
# Hãy đọc README.md để cài bản hỗ trợ GPU (CUDA).

numpy>=1.18.5
opencv-python>=4.1.1
PyYAML>=5.3.1
psutil
tqdm
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
requests>=2.23.0
scipy>=1.4.1
thop  # Thư viện tính toán FLOPs cho YOLOv5
pillow